Move Resource to SDK
