+++ api/lib/opentelemetry.rb
@@ -10,7 +10,6 @@ require 'opentelemetry/context'
 require 'opentelemetry/distributed_context'
 require 'opentelemetry/internal'
 require 'opentelemetry/metrics'
-require 'opentelemetry/resources'
 require 'opentelemetry/trace'
 require 'opentelemetry/version'
