+++ sdk/lib/opentelemetry/sdk/resources.rb
@@ -4,4 +4,4 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-require 'opentelemetry/resources/resource'
+require 'opentelemetry/sdk/resources/resource'
