+++ sdk/lib/opentelemetry/sdk/trace/tracer.rb
@@ -9,6 +9,8 @@ module OpenTelemetry
     module Trace
       # {Tracer} is the SDK implementation of {OpenTelemetry::Trace::Tracer}.
       class Tracer
+        attr_accessor :resource # TODO: should this be read-only? If not, how should exporters know when it is updated?
+
         # Returns a new {Tracer} instance.
         #
         # @return [Tracer]
@@ -17,6 +19,7 @@ module OpenTelemetry
           @registered_span_processors = []
           @mutex = Mutex.new
           @stopped = false
+          @resource = Resources::Resource.create # TODO: should this initialize from env vars instead?
         end
 
         # Attempts to stop all the activity for this {Tracer}. Calls
