+++ sdk/test/opentelemetry/sdk/resources/resource_test.rb
@@ -6,42 +6,40 @@
 
 require 'test_helper'
 
-describe OpenTelemetry::Resources::Resource do
+describe OpenTelemetry::SDK::Resources::Resource do
+  Resource = OpenTelemetry::SDK::Resources::Resource
+
   describe '.new' do
     it 'is private' do
-      -> { OpenTelemetry::Resources::Resource.new('k1' => 'v1') }\
-        .must_raise(NoMethodError)
+      proc { Resource.new('k1' => 'v1') }.must_raise(NoMethodError)
     end
   end
+
   describe '.create' do
     it 'can be initialized with labels' do
       expected_labels = { 'k1' => 'v1', 'k2' => 'v2' }
-      resource = OpenTelemetry::Resources::Resource.create(expected_labels)
+      resource = Resource.create(expected_labels)
       resource.label_enumerator.to_h.must_equal(expected_labels)
     end
 
     it 'can be empty' do
-      resource = OpenTelemetry::Resources::Resource.create
+      resource = Resource.create
       resource.label_enumerator.to_h.must_be_empty
     end
 
     it 'enforces keys are strings' do
-      -> { OpenTelemetry::Resources::Resource.create(k1: 'v1') }\
-        .must_raise(ArgumentError)
+      proc { Resource.create(k1: 'v1') }.must_raise(ArgumentError)
     end
 
     it 'enforces values are strings' do
-      -> { OpenTelemetry::Resources::Resource.create('k1' => :v1) }\
-        .must_raise(ArgumentError)
+      proc { Resource.create('k1' => :v1) }.must_raise(ArgumentError)
     end
   end
 
   describe '#merge' do
     it 'merges two resources into a third' do
-      res1 = OpenTelemetry::Resources::Resource.create('k1' => 'v1',
-                                                       'k2' => 'v2')
-      res2 = OpenTelemetry::Resources::Resource.create('k3' => 'v3',
-                                                       'k4' => 'v4')
+      res1 = Resource.create('k1' => 'v1', 'k2' => 'v2')
+      res2 = Resource.create('k3' => 'v3', 'k4' => 'v4')
       res3 = res1.merge(res2)
 
       res3.label_enumerator.to_h.must_equal('k1' => 'v1', 'k2' => 'v2',
@@ -51,10 +49,8 @@ describe OpenTelemetry::Resources::Resource do
     end
 
     it 'does not overwrite receiver\'s keys when value is non-empty' do
-      res1 = OpenTelemetry::Resources::Resource.create('k1' => 'v1',
-                                                       'k2' => 'v2')
-      res2 = OpenTelemetry::Resources::Resource.create('k2' => '2v2',
-                                                       'k3' => '2v3')
+      res1 = Resource.create('k1' => 'v1', 'k2' => 'v2')
+      res2 = Resource.create('k2' => '2v2', 'k3' => '2v3')
       res3 = res1.merge(res2)
 
       res3.label_enumerator.to_h.must_equal('k1' => 'v1',
@@ -63,10 +59,8 @@ describe OpenTelemetry::Resources::Resource do
     end
 
     it 'overwrites receiver\'s key when value is empty' do
-      res1 = OpenTelemetry::Resources::Resource.create('k1' => 'v1',
-                                                       'k2' => '')
-      res2 = OpenTelemetry::Resources::Resource.create('k2' => '2v2',
-                                                       'k3' => '2v3')
+      res1 = Resource.create('k1' => 'v1', 'k2' => '')
+      res2 = Resource.create('k2' => '2v2', 'k3' => '2v3')
       res3 = res1.merge(res2)
 
       res3.label_enumerator.to_h.must_equal('k1' => 'v1',
